requirements.txt
python-telegram-bot[job-queue]==21.6
requests

